refactoring files utility methods
